add dynamic shape support for part of schedule and refactor ir_schedule
