+++ paddle/cinn/ir/CMakeLists.txt
@@ -27,4 +27,3 @@ add_subdirectory(test)
 add_subdirectory(utils)
 add_subdirectory(schedule)
 add_subdirectory(group_schedule)
-add_subdirectory(dy_schedule)
