--- paddle/cinn/ir/dy_schedule/storage.cc
@@ -1,44 +1,0 @@
-// Copyright (c) 2023 CINN Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "paddle/cinn/common/macros.h"
-#include "paddle/cinn/ir/dy_schedule/ir_schedule.h"
-
-namespace cinn {
-namespace ir {
-
-Expr DyScheduleImpl::CacheRead(const Expr& block,
-                               int read_buffer_index,
-                               const std::string& memory_type) {
-  CINN_NOT_IMPLEMENTED;
-}
-
-Expr DyScheduleImpl::CacheWrite(const Expr& block,
-                                int write_buffer_index,
-                                const std::string& memory_type) {
-  CINN_NOT_IMPLEMENTED;
-}
-
-void DyScheduleImpl::SyncThreads(const Expr& ir_node, bool after_node) {
-  CINN_NOT_IMPLEMENTED;
-}
-
-void DyScheduleImpl::SetBuffer(Expr& block,  // NOLINT
-                               const std::string& memory_type,
-                               bool fixed) {
-  CINN_NOT_IMPLEMENTED;
-}
-
-}  // namespace ir
-}  // namespace cinn
