+++ paddle/cinn/ir/schedule/CMakeLists.txt
@@ -10,6 +10,9 @@ gather_srcs(
   ir_schedule_error.cc
   schedule_desc.cc)
 
+add_subdirectory(impl)
+add_subdirectory(utils)
+
 foreach(header ${schedule_desc_proto_HDRS})
   set(core_proto_includes
       "${core_proto_includes};${header}"
