+++ paddle/cinn/ir/schedule/ir_schedule.cc
@@ -30,7 +30,6 @@
 #include "paddle/cinn/common/dev_info_manager.h"
 #include "paddle/cinn/common/ir_util.h"
 #include "paddle/cinn/common/target.h"
-#include "paddle/cinn/ir/dy_schedule/ir_schedule.h"
 #include "paddle/cinn/ir/ir.h"
 #include "paddle/cinn/ir/ir_analyzer/ir_analyzer.h"
 #include "paddle/cinn/ir/ir_mutator.h"
@@ -38,6 +37,7 @@
 #include "paddle/cinn/ir/ir_visitor.h"
 #include "paddle/cinn/ir/op/ir_operators.h"
 #include "paddle/cinn/ir/schedule/factorize_reduction.h"
+#include "paddle/cinn/ir/schedule/impl/ir_schedule.h"
 #include "paddle/cinn/ir/schedule/ir_schedule_error.h"
 #include "paddle/cinn/ir/schedule/ir_schedule_util.h"
 #include "paddle/cinn/ir/utils/ir_copy.h"
@@ -50,79 +50,6 @@ PD_DECLARE_int32(cinn_error_message_level);
 
 namespace cinn {
 namespace ir {
-
-/**
- * A struct helps to implement static shape Schedule primitives.
- */
-class StScheduleImpl : public ScheduleBase {
- public:
-  StScheduleImpl() = delete;
-  explicit StScheduleImpl(const ModuleExpr& module_expr,
-                          bool debug_flag = false,
-                          utils::ErrorMessageLevel err_msg_level =
-                              utils::ErrorMessageLevel::kGeneral)
-      : ScheduleBase(module_expr, false, err_msg_level) {}
-  explicit StScheduleImpl(ModuleExpr&& module_expr)
-      : ScheduleBase(std::move(module_expr)) {}
-
-  void MergeExprs();
-  bool HasBlock(const std::string& block_name) const;
-  std::vector<Expr> GetLoops(const Expr& block) const;
-  std::vector<Expr> GetLoops(const std::string& block_name) const;
-  std::vector<Expr> GetAllBlocks() const;
-  std::vector<Expr> GetChildBlocks(const Expr& expr) const;
-  Expr GetBlock(const std::string& block_name) const;
-  std::vector<Expr> Split(const Expr& loop, const std::vector<int>& factors);
-  std::vector<Expr> SamplePerfectTile(
-      utils::LinearRandomEngine::StateType* rand_seed,
-      const Expr& loop,
-      int n,
-      int max_innermost_factor);
-  Expr Fuse(const std::vector<Expr>& loops);
-  Expr Fuse(const std::string& block_name, const std::vector<int>& loops_index);
-  Expr Fuse(const Expr& block, const std::vector<int>& loops_index);
-  void ComputeAt(const Expr& block, const Expr& loop, bool keep_unit_loops);
-  void SimpleComputeAt(const Expr& block, const Expr& loop);
-  void ReverseComputeAt(const Expr& block,
-                        const Expr& loop,
-                        bool keep_unit_loops);
-  Expr GetRootBlock(const Expr& expr) const;
-  Expr CacheRead(const Expr& block,
-                 int read_buffer_index,
-                 const std::string& memory_type);
-  Expr CacheWrite(const Expr& block,
-                  int write_buffer_index,
-                  const std::string& memory_type);
-  void SyncThreads(const Expr& ir_node, bool after_node = true);
-  void SetBuffer(Expr& block,  // NOLINT
-                 const std::string& memory_type,
-                 bool fixed = false);
-  Expr Reorder(const std::vector<Expr>& loops);
-  Expr Reorder(const std::string& block_name,
-               const std::vector<int>& loops_index);
-  Expr Reorder(const Expr& block, const std::vector<int>& loops_index);
-  DeviceAPI GetDeviceAPI() const;
-  void MutateForType(const Expr& loop, ForType for_type, int factor = -1);
-  void Parallel(const Expr& loop);
-  void Vectorize(const Expr& loop, int factor);
-  void Unroll(const Expr& loop);
-  void ComputeInline(const Expr& schedule_block);
-  void ReverseComputeInline(const Expr& schedule_block);
-  void Bind(const Expr& loop, const std::string& thread_axis);
-  Expr Rfactor(const Expr& rf_loop, int rf_axis);
-  Expr FactorizeReduction(const Expr& rf_loop, int rf_axis);
-  Expr AddUnitLoop(const Expr& block) const;
-  void Annotate(const Expr& block, const std::string& key, const attr_t& value);
-  void Unannotate(Expr& block, const std::string& key);  // NOLINT
-  void FlattenLoops(const std::vector<Expr>& loops,
-                    const bool force_flat = false);
-  void CopyTransformAndLoopInfo(const Expr& block, const Expr& block_target);
-  void CopyTransformAndLoopInfo(const std::string& block_name,
-                                const std::string& block_target_name);
-  Expr SampleCategorical(utils::LinearRandomEngine::StateType* rand_seed,
-                         const std::vector<int>& candidates,
-                         const std::vector<float>& probs);
-};
 
 std::unique_ptr<ScheduleBase> ScheduleBase::Make(
     const ModuleExpr& module_expr,
@@ -164,229 +91,6 @@ std::unique_ptr<ScheduleBase> ScheduleBase::Make(ModuleExpr&& module_expr,
   catch (const utils::ErrorHandler& err_hanlder) {             \
     CINN_THROW(err_hanlder.FormatErrorMessage(err_msg_level)); \
   }
-
-std::vector<Expr> StScheduleImpl::Split(const Expr& loop,
-                                        const std::vector<int>& factors) {
-  CHECK(loop.As<ir::For>())
-      << "Expr param of Split must be For node! Please check.";
-  auto* for_node = loop.As<ir::For>();
-  CHECK(common::is_zero(for_node->min))
-      << "The For node must start with 0! Please check.";
-  CHECK(for_node->extent.is_constant())
-      << "The For node's extent must be constant! Please check.";
-  int tot_extent = for_node->extent.get_constant();
-
-  VLOG(3) << "Try Split loop from (" << for_node->loop_var->name << ", 0, "
-          << tot_extent << ") to (" << cinn::utils::Join(factors, ", ")
-          << ") at loop:\n"
-          << loop;
-
-  std::vector<int> processed_factors;
-  CINN_IR_SCHEDULE_BEGIN();
-  processed_factors = ValidateFactors(factors, tot_extent, this->module_expr_);
-  CINN_IR_SCHEDULE_END(this->err_msg_level_);
-  int prod_size = std::accumulate(processed_factors.begin(),
-                                  processed_factors.end(),
-                                  1,
-                                  std::multiplies<int>());
-  std::vector<Var> new_loop_vars;
-  Expr substitute_value(0);
-  for (int i = 0; i < processed_factors.size(); ++i) {
-    Var temp_var(common::UniqName(for_node->loop_var->name));
-    substitute_value =
-        Expr(temp_var) + substitute_value * Expr(processed_factors[i]);
-    new_loop_vars.push_back(temp_var);
-  }
-  substitute_value = common::AutoSimplify(substitute_value);
-  Expr new_node = ir::ir_utils::IRCopy(for_node->body);
-  ReplaceExpr(&new_node, {for_node->loop_var}, {substitute_value});
-  std::vector<Expr> splited_loops;
-  splited_loops.resize(processed_factors.size());
-  if (tot_extent < prod_size) {
-    new_node = IfThenElse::Make(LT::Make(substitute_value, for_node->extent),
-                                new_node);
-  }
-  for (int i = processed_factors.size() - 1; i >= 0; i--) {
-    if (!new_node.As<ir::Block>()) new_node = Block::Make({new_node});
-    new_node = For::Make(new_loop_vars[i],
-                         Expr(0),
-                         Expr(processed_factors[i]),
-                         for_node->for_type(),
-                         for_node->device_api,
-                         new_node);
-    splited_loops[i] = new_node;
-  }
-
-  this->Replace(loop, new_node);
-  VLOG(3) << "After Split, ir is:\n" << splited_loops.at(0);
-  return splited_loops;
-}
-
-Expr StScheduleImpl::Fuse(const std::vector<Expr>& loops) {
-  VLOG(3) << "Tring to fuse:\n" << cinn::utils::Join(loops, "\n");
-  std::vector<const ir::For*> for_nodes;
-  std::vector<Var> loop_vars;
-  CHECK(!loops.empty())
-      << "The loops param of Fuse should not be empty! Please check.";
-
-  for (const Expr& it_loop : loops) {
-    CHECK(it_loop.As<ir::For>())
-        << "Expr param of Fuse must be For node! Please check.";
-    if (!for_nodes.empty()) {
-      CHECK(for_nodes.back()->body.As<ir::Block>())
-          << "The body of for node is not Block!";
-      CHECK_EQ(for_nodes.back()->body.As<ir::Block>()->stmts.size(), 1U)
-          << "The Block'size of for node is not 1!";
-      CHECK_EQ(for_nodes.back()->body.As<ir::Block>()->stmts[0], it_loop)
-          << "The For nodes in loops param of Fuse must be adjacent! Please "
-             "check.";
-    }
-    for_nodes.push_back(it_loop.As<ir::For>());
-    loop_vars.push_back(it_loop.As<ir::For>()->loop_var);
-  }
-  std::string suffix;
-  suffix = for_nodes[0]->loop_var->name;
-  int loops_number = for_nodes.size();
-  for (int i = 1; i < loops_number; ++i) {
-    suffix += "_" + for_nodes[i]->loop_var->name;
-  }
-  suffix += "_fused";
-  Var fused_var(suffix);
-  std::vector<Expr> substitute_value;
-  substitute_value.resize(loops_number);
-  Expr fused_expr(fused_var);
-  for (int i = loops_number - 1; i > 0; i--) {
-    substitute_value[i] = Mod::Make(fused_expr, for_nodes[i]->extent);
-    fused_expr = Div::Make(fused_expr, for_nodes[i]->extent);
-  }
-  substitute_value[0] = fused_expr;
-
-  Expr fused_body = ir::ir_utils::IRCopy(for_nodes.back()->body);
-  ReplaceExpr(&fused_body, loop_vars, substitute_value);
-  optim::Simplify(&fused_body);
-  Expr fused_extent(1);
-  for (int i = 0; i < loops_number; ++i) {
-    fused_extent = fused_extent * for_nodes[i]->extent;
-  }
-  fused_extent = common::AutoSimplify(fused_extent);
-
-  if (!fused_body.As<ir::Block>()) fused_body = Block::Make({fused_body});
-  Expr new_stmt = For::Make(fused_var,
-                            Expr(0),
-                            fused_extent,
-                            for_nodes[0]->for_type(),
-                            for_nodes[0]->device_api,
-                            fused_body);
-  this->Replace(loops[0], new_stmt);
-
-  VLOG(3) << "After fuse, ir is:\n" << new_stmt;
-  return new_stmt;
-}
-
-Expr StScheduleImpl::Fuse(const std::string& block_name,
-                          const std::vector<int>& loops_index) {
-  std::vector<Expr> all_loops = this->GetLoops(block_name);
-  std::vector<Expr> loops_expr;
-  loops_expr.reserve(loops_index.size());
-  for (int i = 0; i < loops_index.size(); ++i) {
-    if (i > 0)
-      CHECK_EQ(loops_index[i - 1] + 1, loops_index[i])
-          << "Loops index in Fuse shoule be continuous!";
-  }
-  for (int i : loops_index) {
-    CHECK_LT(i, (int)all_loops.size())
-        << "The loop index in Fuse should be less than total loop's number.";
-    CHECK_GE(i, 0) << "The loop index in Fuse should be >= 0.";
-    loops_expr.emplace_back(all_loops[i]);
-  }
-  return this->Fuse(loops_expr);
-}
-
-Expr StScheduleImpl::Fuse(const Expr& block,
-                          const std::vector<int>& loops_index) {
-  std::vector<Expr> all_loops = this->GetLoops(block);
-  std::vector<Expr> loops_expr;
-  loops_expr.reserve(loops_index.size());
-  for (int i = 0; i < loops_index.size(); ++i) {
-    if (i > 0)
-      CHECK_EQ(loops_index[i - 1] + 1, loops_index[i])
-          << "Loops index in Fuse shoule be continuous!";
-  }
-  for (int i : loops_index) {
-    CHECK_LT(i, (int)all_loops.size())
-        << "The loop index in Fuse should be less than total loop's number.";
-    CHECK_GE(i, 0) << "The loop index in Fuse should be >= 0.";
-    loops_expr.emplace_back(all_loops[i]);
-  }
-  return this->Fuse(loops_expr);
-}
-
-void StScheduleImpl::MutateForType(const Expr& loop,
-                                   ForType for_type,
-                                   int factor) {
-  auto* for_node = loop.As<ir::For>();
-  CHECK(for_node) << "loop param must be For node! Please check.";
-  CHECK(for_node->is_serial())
-      << "loop is not serial, current forloop type is "
-      << static_cast<int>(for_node->for_type()) << ", and it cannot become "
-      << static_cast<int>(for_type);
-  auto loop_copy = ir::ir_utils::IRCopy(loop);
-  auto* new_for_node = loop_copy.As<ir::For>();
-  CHECK(new_for_node);
-  new_for_node->set_for_type(for_type);
-  if (new_for_node->is_vectorized()) {
-    VectorizeInfo vec_info(0, factor);
-    new_for_node->set_vectorize_info(vec_info);
-  } else if (new_for_node->is_binded()) {
-    BindInfo bind_info(for_type, factor, DeviceAPI::GPU);
-    new_for_node->set_bind_info(bind_info);
-  }
-  this->Replace(loop, loop_copy);
-}
-
-void StScheduleImpl::Parallel(const Expr& loop) {
-  MutateForType(loop, ForType::Parallel);
-}
-
-void StScheduleImpl::Vectorize(const Expr& loop, int factor) {
-  CHECK_GT(factor, 0) << "vectorize factor should be more than 0";
-  MutateForType(loop, ForType::Vectorized, factor);
-}
-
-void StScheduleImpl::Unroll(const Expr& loop) {
-  MutateForType(loop, ForType::Unrolled);
-}
-
-void StScheduleImpl::Bind(const Expr& loop, const std::string& thread_axis) {
-#ifdef CINN_WITH_CUDA
-  static std::set<std::string> thread_axes = {"blockIdx.x",
-                                              "blockIdx.y",
-                                              "blockIdx.z",
-                                              "threadIdx.x",
-                                              "threadIdx.y",
-                                              "threadIdx.z"};
-  CHECK(thread_axes.count(thread_axis))
-      << "thread_axis " << thread_axis << " is not supported";
-  int offset = thread_axis.back() - 'x';
-  auto cur_dev_info =
-      common::DevInfoMgr<common::Target::Arch::NVGPU>::GetDevInfo(0);
-  const std::array<int, 3> kMaxBlockDims = cur_dev_info->GetMaxBlockDims();
-  const std::array<int, 3> kMaxGridDims = cur_dev_info->GetMaxGridDims();
-  auto check_offset = [&](const char& c) -> bool {
-    auto extent = loop.As<ir::For>()->extent.as_int32();
-    return extent <= (c == 'b' ? kMaxGridDims[offset] : kMaxBlockDims[offset]);
-  };
-  if (thread_axis[0] == 'b') {
-    CHECK(check_offset(thread_axis[0]))
-        << "Invalid Bind! The extent of loop is out of range on grid size!\n";
-    MutateForType(loop, ForType::GPUBlock, offset);
-  } else {
-    CHECK(check_offset(thread_axis[0]))
-        << "Invalid Bind! The extent of loop is out of range on block size!\n";
-    MutateForType(loop, ForType::GPUThread, offset);
-  }
-#endif
-}
 
 // The struct used to mutate new rfactor forloop and its' schedule block.
 struct RfMutator : public ir::IRMutator<> {
@@ -685,871 +389,6 @@ struct FinalMutator : public ir::IRMutator<> {
   std::vector<Expr> new_rf_indice_;
 };
 
-// The struct used to create all stmts after rfactor transformation.
-struct RfCreater : public ir::IRMutator<> {
- public:
-  RfCreater(const Expr& root, const Expr& rf_loop, const int& rf_axis)
-      : root_(root), rf_loop_(rf_loop), rf_axis_(rf_axis) {}
-  void operator()(Expr* expr) { IRMutator::Visit(expr, expr); }
-
-  Expr CreateRfAllStmts() {
-    auto root_realize = root_.As<ScheduleBlockRealize>();
-    CHECK(root_realize);
-    auto root_block = root_realize->schedule_block.As<ScheduleBlock>();
-    CHECK(root_block);
-    Expr root_loop = ir::ir_utils::IRCopy(root_block->body);
-    if (auto block = root_loop.As<Block>()) {
-      CHECK_EQ(block->stmts.size(), 1U)
-          << "rfactor root should only have one block stmt";
-      root_loop = block->stmts[0];
-    }
-    auto* root_for = root_loop.As<For>();
-    CHECK(root_for);
-    auto rf_for = rf_loop_.As<For>();
-    CHECK(rf_for);
-    // create new rfactor forloops
-    Expr new_rf_forloop = ir::ir_utils::IRCopy(root_loop);
-    RfMutator rf_mutator(rf_loop_, rf_axis_);
-    rf_mutator(&new_rf_forloop);
-    VLOG(3) << "After RfMutator, new rf_forloop is\n" << new_rf_forloop;
-    auto new_rf_tensor = rf_mutator.GetNewRfTensor();
-    // create final write-back forloops
-    Expr final_forloop = ir::ir_utils::IRCopy(root_loop);
-    FinalMutator final_mutator(rf_loop_, rf_axis_, new_rf_tensor);
-    final_mutator(&final_forloop);
-    VLOG(3) << "After FinalMuator, final write-back forloop is\n"
-            << final_forloop;
-    // combine the new created rfactor forloops with the final write-back
-    // forloops and replace
-    root_block->body = Block::Make({new_rf_forloop, final_forloop});
-    return new_rf_tensor;
-  }
-
-  Expr root_;
-  Expr rf_loop_;
-  int rf_axis_;
-};
-
-Expr StScheduleImpl::Rfactor(const Expr& rf_loop, int rf_axis) {
-  CHECKRfactorValidation(rf_loop, rf_axis);
-  // get root ScheduleBlockRealize
-  Expr root = GetRootBlock(rf_loop);
-  // create all stmts after rfactor transformation
-  RfCreater rf_create(root, rf_loop, rf_axis);
-  // return new created rfactor tensor
-  return rf_create.CreateRfAllStmts();
-}
-
-Expr StScheduleImpl::FactorizeReduction(const Expr& rf_loop, int rf_axis) {
-  std::string primitive = "FactorizeReduction";
-  // Get child block of the rf_loop and check.
-  std::vector<Expr> blocks = GetChildBlocks(rf_loop);
-  if (blocks.size() != 1) {
-    std::ostringstream os;
-    os << "The rf_loop is required to have only one child block, but got "
-       << blocks.size() << std::endl;
-    throw IRScheduleErrorHandler(primitive, os.str(), this->module_expr_);
-  }
-  Expr original_block = blocks.at(0);
-  Expr root_block = GetRootBlock(original_block);
-  // TODO(BiynXu): Add CheckReductionBlock()
-
-  // Collect the loops of the block.
-  // Construct a map from loop var names to corresponding loops.
-  std::vector<Expr> original_loops = this->GetLoops(original_block);
-  CHECK_GT(original_loops.size(), 0);
-  VLOG(3) << "before FactorizeReduction, original computational body of the "
-             "reduction is:\n"
-          << original_loops[0];
-  std::map<Var, Expr, CompVar> var2loops;
-  for (const Expr& loop : original_loops) {
-    var2loops[loop.As<For>()->loop_var] = loop;
-  }
-
-  // Get original stmt of reduction update and original store tensor.
-  Expr original_update_body = original_block.As<ir::ScheduleBlockRealize>()
-                                  ->schedule_block.As<ir::ScheduleBlock>()
-                                  ->body;
-  Expr original_update_stmt;
-  CHECK(original_update_body.As<Block>() || original_update_body.As<Store>());
-  if (original_update_body.As<Block>()) {
-    CHECK_EQ(original_update_body.As<Block>()->stmts.size(), 1);
-    original_update_stmt = original_update_body.As<Block>()->stmts[0];
-  } else if (original_update_body.As<Store>()) {
-    original_update_stmt = original_update_body;
-  }
-  Tensor original_tensor =
-      original_update_stmt.As<Store>()->tensor.as_tensor_ref();
-
-  // Create new blocks and loops.
-  Tensor rf_tensor = CreateRFTensor(original_tensor, rf_loop, rf_axis);
-  RFBlockCreater rf_block_creater(original_block,
-                                  original_loops,
-                                  rf_loop,
-                                  original_update_stmt,
-                                  rf_tensor,
-                                  var2loops,
-                                  rf_axis);
-  rf_block_creater.CreateBlock();
-  RBBlockCreater wb_block_creater(original_block,
-                                  original_loops,
-                                  rf_loop,
-                                  original_update_stmt,
-                                  rf_tensor,
-                                  rf_block_creater.rf_tensor_access_indices_,
-                                  rf_block_creater.rf_var_);
-  wb_block_creater.CreateBlock();
-
-  Expr rf_body = rf_block_creater.CreateLoops();
-  Expr wb_body = wb_block_creater.CreateLoops();
-
-  Expr new_computational_body = Block::Make({rf_body, wb_body});
-
-  // Replace and update the AST.
-  this->Replace(original_loops[0], new_computational_body);
-  VLOG(3) << "After FactorizeReduction, new computational body of the "
-             "reduction is:\n"
-          << new_computational_body;
-  return rf_tensor;
-}
-
-struct CacheReadRewriter : public ir::IRMutator<> {
- public:
-  static Expr Rewrite(const Expr& root, CacheBlockInfo* info) {
-    CacheReadRewriter rewriter(root, info);
-    Expr new_root = ir::ir_utils::IRCopy(root);
-    rewriter(&new_root);
-    return new_root;
-  }
-
-  void operator()(Expr* expr) { IRMutator::Visit(expr, expr); }
-
- private:
-  explicit CacheReadRewriter(const Expr& root, CacheBlockInfo* info)
-      : root_(root), info_(info) {}
-
-  void Visit(const ir::Block* expr, Expr* op) override {
-    if (*op == info_->loc_block) {
-      IRMutator::Visit(expr, op);
-      op->As<Block>()->stmts.insert(
-          op->As<Block>()->stmts.begin() + info_->loc_pos, info_->cache_block);
-    } else {
-      IRMutator::Visit(expr, op);
-    }
-  }
-
-  void Visit(const ir::Load* expr, Expr* op) override {
-    if (expr->tensor == Expr(info_->read_tensor)) {
-      IRMutator::Visit(expr, op);
-      op->As<Load>()->tensor = Expr(info_->write_tensor);
-    } else {
-      IRMutator::Visit(expr, op);
-    }
-  }
-
- private:
-  /*! \brief The parent scope of the insertion */
-  const Expr& root_;
-  /*! \brief The info for inserting cache stage */
-  CacheBlockInfo* info_;
-};
-
-struct CacheWriteRewriter : public ir::IRMutator<> {
- public:
-  static Expr Rewrite(const Expr& root, CacheBlockInfo* info) {
-    CacheWriteRewriter rewriter(root, info);
-    Expr new_root = ir::ir_utils::IRCopy(root);
-    rewriter.mutate_cache_block = true;
-    rewriter(&info->cache_block);
-    rewriter.mutate_cache_block = false;
-    rewriter(&new_root);
-    auto find_tensor = ir::ir_utils::CollectIRNodesWithoutTensor(
-        new_root,
-        [&](const Expr* x) {
-          return x->As<Store>() &&
-                 (x->As<Store>()->tensor == Expr(info->read_tensor));
-        },
-        true);
-    if (!find_tensor.empty()) {
-      auto find_store = ir::ir_utils::CollectIRNodesWithoutTensor(
-          (*find_tensor.begin()), [&](const Expr* x) {
-            return x->As<Load>() &&
-                   (x->As<Load>()->tensor == Expr(info->write_tensor));
-          });
-      for (auto load_ir : find_store) {
-        load_ir.As<Load>()->tensor = Expr(info->read_tensor);
-      }
-    }
-    return new_root;
-  }
-
-  void operator()(Expr* expr) { IRMutator::Visit(expr, expr); }
-
- private:
-  explicit CacheWriteRewriter(const Expr& root, CacheBlockInfo* info)
-      : root_(root), info_(info) {}
-
-  void Visit(const ir::Block* expr, Expr* op) override {
-    if (*op == info_->loc_block) {
-      IRMutator::Visit(expr, op);
-      op->As<Block>()->stmts.insert(
-          op->As<Block>()->stmts.begin() + info_->loc_pos, info_->cache_block);
-    } else {
-      IRMutator::Visit(expr, op);
-    }
-  }
-
-  void Visit(const ir::ScheduleBlock* expr, Expr* op) override {
-    if (op->As<ScheduleBlock>()->name == info_->write_tensor->name) {
-      op->As<ScheduleBlock>()->name = info_->read_tensor->name;
-    } else if (op->As<ScheduleBlock>()->name == info_->read_tensor->name) {
-      op->As<ScheduleBlock>()->name = info_->write_tensor->name;
-    }
-    IRMutator::Visit(expr, op);
-  }
-
-  void Visit(const ir::Load* expr, Expr* op) override {
-    IRMutator::Visit(expr, op);
-    if (op->As<Load>()->tensor == Expr(info_->write_tensor) &&
-        mutate_cache_block) {
-      op->As<Load>()->tensor = Expr(info_->read_tensor);
-    } else if (op->As<Load>()->tensor == Expr(info_->read_tensor) &&
-               mutate_cache_block) {
-      op->As<Load>()->tensor = Expr(info_->write_tensor);
-    }
-  }
-
-  void Visit(const ir::Store* expr, Expr* op) override {
-    IRMutator::Visit(expr, op);
-    if (op->As<Store>()->tensor == Expr(info_->write_tensor)) {
-      op->As<Store>()->tensor = Expr(info_->read_tensor);
-    } else if (op->As<Store>()->tensor == Expr(info_->read_tensor) &&
-               mutate_cache_block) {
-      op->As<Store>()->tensor = Expr(info_->write_tensor);
-    }
-  }
-
- private:
-  /*! \brief The parent scope of the insertion */
-  const Expr& root_;
-  /*! \brief The info for inserting cache stage */
-  CacheBlockInfo* info_;
-  /*! \brief Are we mutating the cache tensor's block */
-  bool mutate_cache_block{true};
-};
-
-//! Visit all ScheduleBlock and change its body to ir::Block if it is not.
-struct ChangeBodyToBlock : public ir::IRMutator<> {
- public:
-  static void Change(Expr* expr) {
-    ChangeBodyToBlock mutator;
-    mutator(expr);
-  }
-
-  void operator()(Expr* expr) { IRMutator::Visit(expr, expr); }
-
- private:
-  void Visit(const ir::ScheduleBlock* expr, Expr* op) override {
-    if (!op->As<ScheduleBlock>()->body.As<Block>()) {
-      op->As<ScheduleBlock>()->body =
-          Block::Make({op->As<ScheduleBlock>()->body});
-    }
-    IRMutator::Visit(expr, op);
-  }
-};
-
-DeviceAPI StScheduleImpl::GetDeviceAPI() const {
-  auto exprs = this->GetModule().GetExprs();
-  return analyzer::GetDeviceAPI(exprs);
-}
-
-Expr StScheduleImpl::CacheRead(const Expr& block,
-                               int read_tensor_index,
-                               const std::string& memory_type) {
-  CHECK(block.As<ScheduleBlockRealize>());
-  auto root = GetRootBlock(block);
-  ChangeBodyToBlock::Change(&root);
-  Expr read_expr = GetNthAccessExpr(block, read_tensor_index, false);
-  CHECK(read_expr.As<ir::Load>());
-  auto tensor_indices = read_expr.As<ir::Load>()->indices;
-  CacheBlockInfo info;
-  info.read_tensor = read_expr.As<ir::Load>()->tensor.as_tensor_ref();
-  info.write_tensor = MakeCacheTensor(info.read_tensor, memory_type);
-  info.alloc = info.write_tensor;
-
-  auto read_ranges =
-      CalculateTensorRegions(block, tensor_indices, info.read_tensor, root);
-  auto new_block =
-      MakeCacheBlock(read_ranges, &info, memory_type, this->GetDeviceAPI());
-  FindInsertionPoint(root, &info, false);
-  auto new_root = CacheReadRewriter::Rewrite(root, &info);
-  this->Replace(
-      root.As<ScheduleBlockRealize>()->schedule_block.As<ScheduleBlock>()->body,
-      new_root.As<ScheduleBlockRealize>()
-          ->schedule_block.As<ScheduleBlock>()
-          ->body);
-  return new_block;
-}
-
-Expr StScheduleImpl::CacheWrite(const Expr& block,
-                                int write_buffer_index,
-                                const std::string& memory_type) {
-  CHECK(block.As<ScheduleBlockRealize>());
-  auto root = GetRootBlock(block);
-  ChangeBodyToBlock::Change(&root);
-  Expr write_expr = GetNthAccessExpr(block, write_buffer_index, true);
-  CHECK(write_expr.As<ir::Store>());
-  Tensor write_tensor = write_expr.As<ir::Store>()->tensor.as_tensor_ref();
-  auto tensor_indices = write_expr.As<ir::Store>()->indices;
-  CacheBlockInfo info;
-  info.read_tensor = MakeCacheTensor(write_tensor, memory_type);
-  info.write_tensor = write_tensor;
-  info.alloc = info.read_tensor;
-  auto write_ranges =
-      CalculateTensorRegions(block, tensor_indices, info.write_tensor, root);
-  auto new_block =
-      MakeCacheBlock(write_ranges, &info, memory_type, this->GetDeviceAPI());
-  FindInsertionPoint(root, &info, true);
-
-  auto new_root = CacheWriteRewriter::Rewrite(root, &info);
-  this->Replace(
-      root.As<ScheduleBlockRealize>()->schedule_block.As<ScheduleBlock>()->body,
-      new_root.As<ScheduleBlockRealize>()
-          ->schedule_block.As<ScheduleBlock>()
-          ->body);
-
-  auto find_cache_block = ir::ir_utils::CollectIRNodesWithoutTensor(
-      root,
-      [&](const Expr* x) {
-        return x->As<ir::ScheduleBlockRealize>() &&
-               !x->As<ir::ScheduleBlockRealize>()->iter_values.empty() &&
-               GetTensor(*x)->name == info.read_tensor->name;
-      },
-      true);
-
-  CHECK(info.write_tensor->buffer.defined());
-
-  // Replace buffer
-  auto all_tensors =
-      ir::ir_utils::CollectIRNodesWithoutTensor(root, [&](const Expr* x) {
-        return x->as_tensor() && x->as_tensor()->buffer.defined();
-      });
-
-  for (auto i : all_tensors) {
-    if (i.as_tensor()->name != info.write_tensor->name &&
-        i.as_tensor()->buffer.defined() &&
-        i.as_tensor()->buffer->name == info.write_tensor->buffer->name) {
-      i.as_tensor()->Bind(info.read_tensor->buffer);
-    }
-  }
-
-  CHECK_EQ(find_cache_block.size(), 1U);
-
-  return *find_cache_block.begin();
-}
-
-struct InsertExpr : public ir::IRMutator<> {
- public:
-  static void Insert(const Expr& ir_node,
-                     const Expr& insert_node,
-                     bool after_node,
-                     Expr* expr) {
-    InsertExpr mutator(ir_node, insert_node, after_node);
-    mutator(expr);
-  }
-
-  void operator()(Expr* expr) { IRMutator::Visit(expr, expr); }
-
- private:
-  explicit InsertExpr(const Expr& ir_node,
-                      const Expr& insert_node,
-                      bool after_node)
-      : ir_node_(ir_node), insert_node_(insert_node), after_node_(after_node) {}
-
-  void Visit(const ir::Block* expr, Expr* op) override {
-    for (int i = 0; i < expr->stmts.size(); i++) {
-      if (expr->stmts[i] == ir_node_) {
-        if (after_node_) {
-          op->As<ir::Block>()->stmts.insert(
-              op->As<ir::Block>()->stmts.begin() + i + 1, insert_node_);
-        } else {
-          op->As<ir::Block>()->stmts.insert(
-              op->As<ir::Block>()->stmts.begin() + i, insert_node_);
-        }
-        return;
-      }
-    }
-    IRMutator::Visit(expr, op);
-  }
-
-  void Visit(const ir::For* expr, Expr* op) override {
-    if (expr->body == ir_node_) {
-      if (after_node_)
-        op->As<ir::For>()->body =
-            ir::Block::Make({op->As<ir::For>()->body, insert_node_});
-      else
-        op->As<ir::For>()->body =
-            ir::Block::Make({insert_node_, op->As<ir::For>()->body});
-      return;
-    }
-    IRMutator::Visit(expr, op);
-  }
-
- private:
-  const Expr& ir_node_;
-  const Expr& insert_node_;
-  bool after_node_;
-};
-
-void StScheduleImpl::SyncThreads(const Expr& ir_node, bool after_node) {
-  CHECK(ir_node.As<ScheduleBlockRealize>() || ir_node.As<ir::For>());
-  auto root = GetRootBlock(ir_node);
-  ChangeBodyToBlock::Change(&root);
-  Expr sync_threads = runtime::IntrinsicCall(Void(), "__syncthreads", {});
-  InsertExpr::Insert(ir_node, sync_threads, after_node, &root);
-  return;
-}
-
-Expr StScheduleImpl::Reorder(const std::vector<Expr>& loops) {
-  if (loops.size() <= 1) {
-    return Expr{nullptr};
-  }
-  VLOG(4) << "Before Reorder, ir is:\n" << loops[0];
-
-  std::set<Expr, CompExpr> loop_set = CollectLoopsToSet(loops);
-  auto boundary = GetBoundaryOfReorderRange(loop_set);
-  Expr top = boundary.first;
-  Expr bottom = boundary.second;
-  std::vector<Expr> chain = GetLoopsInRange(top, bottom);
-  std::vector<Expr> if_nodes = GetIfThenElseInRange(top, bottom);
-  Expr new_loop = ConstructNewLoopChain(chain, loops, loop_set, if_nodes);
-  this->Replace(top, new_loop);
-
-  VLOG(4) << "After Reorder, ir is:\n" << new_loop;
-  return new_loop;
-}
-
-Expr StScheduleImpl::Reorder(const std::string& block_name,
-                             const std::vector<int>& loops_index) {
-  std::vector<Expr> all_loops = this->GetLoops(block_name);
-  std::vector<Expr> loops_expr;
-  loops_expr.reserve(loops_index.size());
-  for (int i : loops_index) {
-    CHECK_LT(i, (int)all_loops.size())
-        << "The loop index in Reorder should be less than total loop's number.";
-    CHECK_GE(i, 0) << "The loop index in Reorder should be >= 0.";
-    loops_expr.emplace_back(all_loops[i]);
-  }
-  return this->Reorder(loops_expr);
-}
-
-Expr StScheduleImpl::Reorder(const Expr& block,
-                             const std::vector<int>& loops_index) {
-  std::vector<Expr> all_loops = this->GetLoops(block);
-  std::vector<Expr> loops_expr;
-  loops_expr.reserve(loops_index.size());
-  for (int i : loops_index) {
-    CHECK_LT(i, (int)all_loops.size())
-        << "The loop index in Reorder should be less than total loop's number.";
-    CHECK_GE(i, 0) << "The loop index in Reorder should be >= 0.";
-    loops_expr.emplace_back(all_loops[i]);
-  }
-  return this->Reorder(loops_expr);
-}
-
-Expr StScheduleImpl::GetRootBlock(const Expr& expr) const {
-  auto exprs = this->GetModule().GetExprs();
-  return analyzer::GetRootBlock(exprs, expr);
-}
-
-// The struct used to reconstruct the new For node to replace the old For node.
-struct LoopReconstructor : public ir::IRMutator<> {
- public:
-  explicit LoopReconstructor(const Expr& root,
-                             const Expr& block,
-                             const Expr& loop)
-      : root_(root), block_(block), loop_(loop) {}
-
-  void operator()(Expr* expr) { IRMutator::Visit(expr, expr); }
-
-  /* \param inserted_pos The position index of the new_loop_ body `stmts` to be
-   * inserted:
-   *        - `index = -1` means inserted into the tail
-   *        - otherwise, it should be a index between [0, stmts size)
-   */
-  std::string MakeNewLoop(const std::vector<IterRange>& iter_ranges,
-                          bool keep_unit_loops,
-                          int inserted_pos = -1) {
-    int n_iters = iter_ranges.size();
-    std::vector<Var> loop_vars;
-    std::vector<Expr> loop_extents;
-    std::vector<Expr> iter_values;
-    loop_vars.reserve(n_iters);
-    loop_extents.reserve(n_iters);
-    iter_values.reserve(n_iters);
-    std::vector<std::string> new_var_names;
-    for (int i = 0; i < n_iters; ++i) {
-      const auto& range = iter_ranges[i];
-      if (keep_unit_loops || range.extent != Expr(1)) {
-        std::string var_name =
-            common::UniqName("ax" + std::to_string(loop_vars.size()));
-        new_var_names.push_back(var_name);
-        Var var(var_name, Int(32));
-        loop_vars.push_back(var);
-        loop_extents.push_back(range.extent);
-        iter_values.push_back(common::AutoSimplify(range.min) + var);
-      } else {
-        iter_values.push_back(common::AutoSimplify(range.min));
-      }
-    }
-    auto schedule_block_node =
-        block_.As<ir::ScheduleBlockRealize>()->schedule_block;
-    new_block_ = ScheduleBlockRealize::Make(std::move(iter_values),
-                                            std::move(schedule_block_node));
-    Expr loop_body = new_block_;
-    for (int i = static_cast<int>(loop_vars.size()) - 1; i >= 0; --i) {
-      auto loop_var = loop_vars[i];
-      auto loop_extent = loop_extents[i];
-      if (!loop_body.As<ir::Block>()) loop_body = Block::Make({loop_body});
-      loop_body = For::Make(loop_var,
-                            Expr(0),
-                            loop_extent,
-                            ForType::Serial,
-                            loop_.As<ir::For>()->device_api,
-                            std::move(loop_body));
-    }
-    new_loop_ = ir::ir_utils::IRCopy(loop_);
-
-    // Replace the copied Tensor object with the original Tensor object,
-    // to ensure that the same Tensor in a AST is the same object.
-    std::unordered_map<std::string, ir::Expr> tensors_map;
-    ir::ir_utils::CollectIRNodesWithoutTensor(
-        loop_, [&tensors_map](const Expr* x) {
-          if (x->as_tensor()) {
-            tensors_map.insert({x->as_tensor()->name, *x});
-            return true;
-          }
-          return false;
-        });
-    auto find_store = ir::ir_utils::CollectIRNodesWithoutTensor(
-        new_loop_, [](const Expr* x) { return x->As<ir::Store>(); });
-    for (auto store : find_store) {
-      store.As<ir::Store>()->tensor =
-          tensors_map.at(store.As<ir::Store>()->tensor.as_tensor()->name);
-    }
-    auto find_load = ir::ir_utils::CollectIRNodesWithoutTensor(
-        new_loop_, [](const Expr* x) { return x->As<ir::Load>(); });
-    for (auto load : find_load) {
-      load.As<ir::Load>()->tensor =
-          tensors_map.at(load.As<ir::Load>()->tensor.as_tensor()->name);
-    }
-
-    InsertBlock(new_loop_, loop_body, inserted_pos);
-    return utils::Join(new_var_names, ",");
-  }
-
- public:
-  /*! \brief The root block */
-  Expr root_;
-  /*! \brief The given block to be moved */
-  Expr block_;
-  /*! \brief The given loop the block and its loop nest to be put under */
-  Expr loop_;
-  /*! \brief The new loop to replace the original loop */
-  Expr new_loop_{nullptr};
-  /*! \brief The new block realize to the moved block */
-  Expr new_block_{nullptr};
-  /*! \brief The plan to remove the given block by replacing this loop/block in
-   * the AST */
-  Expr source_expr{nullptr};
-  /*! \brief The plan to remove the given block by replacing to this loop/block
-   * in the AST */
-  Expr target_expr{nullptr};
-};
-
-struct FixLocalBufferSize : public ir::IRMutator<> {
- public:
-  explicit FixLocalBufferSize(const std::string& tensor_name)
-      : tensor_name_(tensor_name) {}
-
-  void operator()(Expr* expr) { IRMutator::Visit(expr, expr); }
-
- private:
-  void Visit(const ir::Store* expr, Expr* op) override {
-    if (op->As<Store>()->tensor.As<_Tensor_>()->name == tensor_name_) {
-      op->As<Store>()->tensor.As<_Tensor_>()->shape = {Expr(1)};
-      op->As<Store>()->tensor.As<_Tensor_>()->domain = {Expr(1)};
-      op->As<Store>()->tensor.As<_Tensor_>()->buffer->shape = {Expr(1)};
-      op->As<Store>()->indices = {Expr(0)};
-    }
-    IRMutator::Visit(expr, op);
-  }
-
-  void Visit(const ir::Load* expr, Expr* op) override {
-    if (op->As<Load>()->tensor.As<_Tensor_>()->name == tensor_name_) {
-      op->As<Load>()->tensor.As<_Tensor_>()->shape = {Expr(1)};
-      op->As<Load>()->tensor.As<_Tensor_>()->domain = {Expr(1)};
-      op->As<Load>()->tensor.As<_Tensor_>()->buffer->shape = {Expr(1)};
-      op->As<Load>()->indices = {Expr(0)};
-    }
-    IRMutator::Visit(expr, op);
-  }
-  std::string tensor_name_;
-};
-
-void StScheduleImpl::SetBuffer(Expr& block,
-                               const std::string& memory_type,
-                               bool fixed) {
-  CHECK(block.As<ir::ScheduleBlockRealize>());
-  auto find_tensor = ir::ir_utils::CollectIRNodesWithoutTensor(
-      block, [&](const Expr* x) { return x->As<ir::Store>(); }, true);
-  CHECK_EQ(find_tensor.size(), 1U)
-      << "One block should only have one Store node!(except for root block)";
-  auto& tensor = (*find_tensor.begin()).As<ir::Store>()->tensor;
-  tensor.as_tensor_ref()->WithBuffer(
-      memory_type, "_" + tensor.as_tensor_ref()->name + "_temp_buffer");
-
-  auto exprs = this->GetModule().GetExprs();
-  for (auto& it_expr : exprs) {
-    auto find_tensor =
-        ir::ir_utils::CollectIRNodesWithoutTensor(it_expr, [&](const Expr* x) {
-          return x->as_tensor() &&
-                 (x->as_tensor()->name == tensor.as_tensor_ref()->name ||
-                  x->as_tensor()->name ==
-                      tensor.as_tensor_ref()->name + "__reduce_init");
-        });
-    for (auto& t : find_tensor) {
-      CHECK(t.as_tensor());
-      t.as_tensor_ref()->Bind(tensor.as_tensor_ref()->buffer);
-    }
-  }
-
-  // if buffer type == "local"
-  if (memory_type == "local" && fixed) {
-    FixLocalBufferSize mutator(block.As<ir::ScheduleBlockRealize>()
-                                   ->schedule_block.As<ir::ScheduleBlock>()
-                                   ->name);
-    auto root = GetRootBlock(block);
-    mutator(&root);
-  }
-}
-
-void StScheduleImpl::MergeExprs() {
-  auto exprs = this->GetModule().GetExprs();
-  if (exprs.size() == 1U) return;
-  CHECK(exprs[0].As<ir::Block>());
-  CHECK_EQ(exprs[0].As<ir::Block>()->stmts.size(), 1U);
-  CHECK(exprs[0].As<ir::Block>()->stmts[0].As<ir::ScheduleBlockRealize>());
-  CHECK(exprs[0]
-            .As<ir::Block>()
-            ->stmts[0]
-            .As<ir::ScheduleBlockRealize>()
-            ->schedule_block.As<ir::ScheduleBlock>());
-  std::vector<Expr> merged_block;
-  merged_block.push_back(exprs[0]
-                             .As<ir::Block>()
-                             ->stmts[0]
-                             .As<ir::ScheduleBlockRealize>()
-                             ->schedule_block.As<ir::ScheduleBlock>()
-                             ->body);
-  VLOG(3) << "Before merging, exprs[0] is : " << exprs[0];
-  for (int i = 1; i < exprs.size(); ++i) {
-    auto root_block = ir::ir_utils::CollectIRNodesWithoutTensor(
-        exprs[i],
-        [&](const Expr* x) {
-          return x->As<ir::ScheduleBlockRealize>() &&
-                 x->As<ir::ScheduleBlockRealize>()->iter_values.empty();
-        },
-        true);
-    CHECK_EQ(root_block.size(), 1U);
-    for (auto& it_block : root_block) {
-      auto& block_body = it_block.As<ir::ScheduleBlockRealize>()
-                             ->schedule_block.As<ir::ScheduleBlock>()
-                             ->body;
-      merged_block.push_back(block_body);
-    }
-  }
-  for (auto& block : merged_block) {
-    VLOG(3) << "in merged_block, it has " << block;
-  }
-  auto merged_expr = ir::Block::Make(merged_block);
-  exprs[0]
-      .As<ir::Block>()
-      ->stmts[0]
-      .As<ir::ScheduleBlockRealize>()
-      ->schedule_block.As<ir::ScheduleBlock>()
-      ->body = merged_expr;
-  VLOG(3) << "After merging, exprs[0] is : " << exprs[0];
-  exprs.erase(exprs.begin() + 1, exprs.end());
-  this->SetExprs(exprs);
-}
-
-void StScheduleImpl::ComputeAt(const Expr& block,
-                               const Expr& loop,
-                               bool keep_unit_loops) {
-  CHECK(block.As<ir::ScheduleBlockRealize>());
-  CHECK(loop.As<ir::For>());
-  Expr root = this->GetRootBlock(block);
-
-  VLOG(3) << "Begin ComputeAt of loop:\n" << loop << "\nat block:\n" << root;
-
-  auto producers = GetProducers(block, root);
-  auto consumers = GetConsumers(block, root);
-  CheckComputeAtValidation(block, loop, root);
-  LoopReconstructor reconstructor(root, block, loop);
-  LeafBlockRemovalPlan remove_plan(
-      block, &reconstructor.source_expr, &reconstructor.target_expr);
-  remove_plan(&root);
-  auto iter_ranges = CalculateRequiredRegions(block, loop, root, consumers);
-  std::string new_var_names =
-      reconstructor.MakeNewLoop(iter_ranges, keep_unit_loops, 0);
-  auto sch_block_expr = block.As<ir::ScheduleBlockRealize>()->schedule_block;
-  sch_block_expr.As<ir::ScheduleBlock>()->attrs.emplace(
-      ir::attr::compute_at_extra_var, new_var_names);
-  this->Replace(reconstructor.source_expr, reconstructor.target_expr);
-  this->Replace(reconstructor.loop_, reconstructor.new_loop_);
-
-  VLOG(3) << "After SimpleComputeAt, ir is:\n" << reconstructor.new_loop_;
-}
-
-void StScheduleImpl::SimpleComputeAt(const Expr& block, const Expr& loop) {
-  CHECK(block.As<ir::ScheduleBlockRealize>());
-  CHECK(loop.As<ir::For>());
-  std::vector<Expr> block_loops = this->GetLoops(block);
-  Expr root = this->GetRootBlock(block);
-  auto loops = GetLoopsOfExpr(loop, root);
-
-  VLOG(3) << "Begin SimpleComputeAt of loop:\n"
-          << loop << "\nat block:\n"
-          << root;
-
-  auto this_loop = loop;
-  auto block_name = GetTensor(block)->name;
-  auto this_block = block;
-  if (GetLoopExtent(loops[0]) == 1 && GetLoopExtent(block_loops[0]) != 1) {
-    this->Split(block_loops[0], {1, -1});
-    this_block = this->GetBlock(block_name);
-  } else if (GetLoopExtent(loops[0]) != 1 &&
-             GetLoopExtent(block_loops[0]) == 1) {
-    auto splited = this->Split(loops[0], {1, -1});
-    this_loop = splited[1];
-  }
-
-  block_loops = this->GetLoops(this_block);
-  root = this->GetRootBlock(this_block);
-  loops = GetLoopsOfExpr(this_loop, root);
-
-  CHECK_LE(loops.size(), block_loops.size());
-
-  std::vector<Var> replaced_var;
-  std::vector<Expr> substitute_expr;
-  for (int i = 0; i < loops.size(); ++i) {
-    CHECK_EQ(GetLoopExtent(loops[i]), GetLoopExtent(block_loops[i]));
-    if (block_loops[i].As<ir::For>()->bind_info().valid() &&
-        !loops[i].As<ir::For>()->bind_info().valid()) {
-      loops[i].As<ir::For>()->set_bind_info(
-          block_loops[i].As<ir::For>()->bind_info());
-    }
-    replaced_var.push_back(block_loops[i].As<ir::For>()->loop_var);
-    substitute_expr.push_back(Expr(loops[i].As<ir::For>()->loop_var));
-  }
-
-  Expr result = loops.size() < block_loops.size()
-                    ? ir::ir_utils::IRCopy(block_loops[loops.size()])
-                    : ir::ir_utils::IRCopy(this_block);
-  Expr new_loop = ir::ir_utils::IRCopy(this_loop);
-
-  // Get the body of block_loop under the same loops
-  auto body = block_loops.at(loops.size() - 1).As<ir::For>()->body;
-  // collect if
-  auto if_checker = [](const Expr* x) { return x->As<ir::IfThenElse>(); };
-  auto if_set = ir::ir_utils::CollectIRNodesWithoutTensor(body, if_checker);
-  for (auto if_expr : if_set) {
-    auto checker = [block_name](const Expr* x) {
-      return x->As<ir::ScheduleBlockRealize>() &&
-             x->As<ir::ScheduleBlockRealize>()
-                     ->schedule_block.As<ScheduleBlock>()
-                     ->name == block_name;
-    };
-    if (ir::ir_utils::CollectIRNodesWithoutTensor(if_expr, checker, true)
-            .size() > 0) {
-      result =
-          IfThenElse::Make(if_expr.As<ir::IfThenElse>()->condition, result);
-      break;
-    }
-  }
-
-  ReplaceExpr(&result, replaced_var, substitute_expr);
-  // When there are two identical IfThenElse
-  if (new_loop.As<ir::For>() && new_loop.As<ir::For>()->body.As<ir::Block>() &&
-      new_loop.As<ir::For>()
-          ->body.As<ir::Block>()
-          ->stmts[0]
-          .As<ir::IfThenElse>()) {
-    auto if_then_else = new_loop.As<ir::For>()->body.As<ir::Block>()->stmts[0];
-    if (result.As<ir::IfThenElse>() &&
-        if_then_else.As<ir::IfThenElse>()->condition ==
-            result.As<ir::IfThenElse>()->condition) {
-      new_loop.As<ir::For>()
-          ->body.As<ir::Block>()
-          ->stmts[0]
-          .As<ir::IfThenElse>()
-          ->true_case = ir::Block::Make({result.As<ir::IfThenElse>()->true_case,
-                                         new_loop.As<ir::For>()
-                                             ->body.As<ir::Block>()
-                                             ->stmts[0]
-                                             .As<ir::IfThenElse>()
-                                             ->true_case});
-    } else {
-      std::vector<ir::Expr>::iterator pos =
-          new_loop.As<ir::For>()->body.As<ir::Block>()->stmts.begin();
-      new_loop.As<ir::For>()->body.As<ir::Block>()->stmts.insert(pos, result);
-    }
-  } else {
-    new_loop.As<ir::For>()->body =
-        ir::Block::Make({result, new_loop.As<ir::For>()->body});
-  }
-
-  Expr source_expr{nullptr};
-  Expr target_expr{nullptr};
-
-  LeafBlockRemovalPlan remove_plan(
-      result.As<ir::For>() ? block_loops[loops.size()] : this_block,
-      &source_expr,
-      &target_expr);
-  remove_plan(&root);
-
-  this->Replace(source_expr, target_expr);
-  this->Replace(this_loop, new_loop);
-
-  VLOG(3) << "After SimpleComputeAt, ir is:\n" << new_loop;
-}
-
-void StScheduleImpl::ReverseComputeAt(const Expr& block,
-                                      const Expr& loop,
-                                      bool keep_unit_loops) {
-  CHECK(block.As<ir::ScheduleBlockRealize>());
-  CHECK(loop.As<ir::For>());
-  Expr root = this->GetRootBlock(block);
-  auto producers = GetProducers(block, root);
-  auto consumers = GetConsumers(block, root);
-  CheckComputeAtValidation(block, loop, root);
-  LoopReconstructor reconstructor(root, block, loop);
-  LeafBlockRemovalPlan remove_plan(
-      block, &reconstructor.source_expr, &reconstructor.target_expr);
-  remove_plan(&root);
-  auto iter_ranges =
-      CalculateRequiredRegions(block, loop, root, producers, false);
-  std::string new_var_names =
-      reconstructor.MakeNewLoop(iter_ranges, keep_unit_loops, -1);
-  auto sch_block_expr = block.As<ir::ScheduleBlockRealize>()->schedule_block;
-  sch_block_expr.As<ir::ScheduleBlock>()->attrs.emplace(
-      ir::attr::reverse_compute_at_extra_var, new_var_names);
-  this->Replace(reconstructor.source_expr, reconstructor.target_expr);
-  this->Replace(reconstructor.loop_, reconstructor.new_loop_);
-  return;
-}
-
 void BaseInliner::operator()(Expr* expr) {
   IRMutator::Visit(&tgt_stmt, &tgt_stmt);
   IRMutator::Visit(expr, expr);
@@ -1637,20 +476,6 @@ Expr ComputeInliner::ReplaceInlinedTensor(Expr* load) {
   return value_copy;
 }
 
-void StScheduleImpl::ComputeInline(const Expr& schedule_block) {
-  CHECK(schedule_block.As<ir::ScheduleBlockRealize>());
-  Expr root = this->GetRootBlock(schedule_block);
-  Expr store = CheckComputeInlineValidationAndGetStore(schedule_block, root);
-  ComputeInliner inliner(store.As<ir::Store>()->tensor.as_tensor_ref(), store);
-  CHECK(inliner.BodyPatternAllowInline());
-  // Create a plan that removes the block to be inlined
-  LeafBlockRemovalPlan remove_plan(
-      schedule_block, &inliner.src_stmt, &inliner.tgt_stmt);
-  remove_plan(&root);
-  inliner(&root);
-  return;
-}
-
 bool ComputeInlineChecker::Check() {
   Expr root = ir_schedule_.GetRootBlock(block_);
   store_ = CheckComputeInlineValidationAndGetStore(block_, root);
@@ -1728,27 +553,6 @@ Expr ReverseComputeInliner::ReplaceTargetTensor(Expr* store) {
   return value_copy;
 }
 
-void StScheduleImpl::ReverseComputeInline(const Expr& schedule_block) {
-  Expr root = this->GetRootBlock(schedule_block);
-  auto exprs =
-      CheckReverseComputeInlineValidationAndGetExprs(schedule_block, root);
-  Expr inlined_load = std::get<0>(exprs);
-  Expr inlined_store = std::get<1>(exprs);
-  Expr target_store = std::get<2>(exprs);
-  ReverseComputeInliner inliner(
-      inlined_store.As<ir::Store>()->tensor.as_tensor_ref(),
-      inlined_store,
-      inlined_load,
-      target_store);
-  CHECK(inliner.BodyPatternAllowInline());
-  // Create a plan that removes the block to be inlined
-  LeafBlockRemovalPlan remove_plan(
-      schedule_block, &inliner.src_stmt, &inliner.tgt_stmt);
-  remove_plan(&root);
-  inliner(&root);
-  inliner(&root);
-}
-
 struct FindBlockParent : public ir::IRMutator<> {
  public:
   explicit FindBlockParent(const std::string& block_name)
@@ -1803,423 +607,6 @@ struct FindBlockParent : public ir::IRMutator<> {
  public:
   ir::Expr* target_{nullptr};
 };
-
-Expr StScheduleImpl::AddUnitLoop(const Expr& block) const {
-  auto exprs = module_expr_.GetExprs();
-  return analyzer::AddUnitLoop(exprs, block);
-}
-
-std::vector<Expr> StScheduleImpl::GetLoops(const Expr& block) const {
-  auto exprs = module_expr_.GetExprs();
-  return analyzer::GetLoops(exprs, block);
-}
-
-std::vector<Expr> StScheduleImpl::GetLoops(
-    const std::string& block_name) const {
-  auto exprs = module_expr_.GetExprs();
-  return analyzer::GetLoops(exprs, block_name);
-}
-
-std::vector<Expr> StScheduleImpl::GetAllBlocks() const {
-  auto exprs = module_expr_.GetExprs();
-  return analyzer::GetAllBlocks(exprs);
-}
-
-std::vector<Expr> StScheduleImpl::GetChildBlocks(const Expr& expr) const {
-  return analyzer::GetChildBlocks(expr);
-}
-
-bool StScheduleImpl::HasBlock(const std::string& block_name) const {
-  auto exprs = module_expr_.GetExprs();
-  return analyzer::HasBlock(exprs, block_name);
-}
-
-Expr StScheduleImpl::GetBlock(const std::string& block_name) const {
-  auto exprs = module_expr_.GetExprs();
-  return analyzer::GetBlock(exprs, block_name);
-}
-
-void StScheduleImpl::Annotate(const Expr& block,
-                              const std::string& key,
-                              const attr_t& value) {
-  CHECK(block.As<ir::ScheduleBlockRealize>());
-  CHECK(block.As<ir::ScheduleBlockRealize>()
-            ->schedule_block.As<ir::ScheduleBlock>());
-  auto copied_block = ir::ir_utils::IRCopy(block);
-  auto* schedule_block = copied_block.As<ir::ScheduleBlockRealize>()
-                             ->schedule_block.As<ir::ScheduleBlock>();
-  schedule_block->attrs.emplace(key, value);
-  this->Replace(block, copied_block);
-}
-
-void StScheduleImpl::Unannotate(Expr& block, const std::string& ann_key) {
-  CHECK(block.As<ir::ScheduleBlockRealize>());
-  CHECK(block.As<ir::ScheduleBlockRealize>()
-            ->schedule_block.As<ir::ScheduleBlock>());
-  auto* schedule_block = block.As<ir::ScheduleBlockRealize>()
-                             ->schedule_block.As<ir::ScheduleBlock>();
-  if (schedule_block->attrs.count(ann_key)) {
-    schedule_block->attrs.erase(ann_key);
-  } else {
-    LOG(WARNING) << "Can't find annotation with key: " << ann_key;
-    return;
-  }
-}
-
-void StScheduleImpl::FlattenLoops(const std::vector<Expr>& loops,
-                                  const bool flat_tensor) {
-  CHECK_GT(loops.size(), 0) << "Loops can't be empty!";
-  VLOG(4) << "Before FlattenLoops, ir is:\n" << loops[0];
-  // compute loop
-  int extent = 1;
-  std::vector<int> strides;
-  std::vector<ir::Var> loop_vars(loops.size());
-  for (int idx = loops.size() - 1; idx >= 0; --idx) {
-    strides.insert(strides.begin(), extent);
-    extent *= loops[idx].As<ir::For>()->extent.as_int32();
-    loop_vars[idx] = loops[idx].As<ir::For>()->loop_var;
-  }
-  CHECK_EQ(loops.size(), strides.size());
-
-  // create new loop.
-  auto last = loops.back().As<ir::For>();
-  auto var = ir::Var("flat_i");
-  auto _var = ir::Var("_flat_i");
-  auto loop = ir::For::Make(var,
-                            ir::Expr(0),
-                            ir::Expr(extent),
-                            last->for_type(),
-                            last->device_api,
-                            last->body);
-
-  // map loop var to old loop var.
-  auto _iter = ir::Expr(_var);
-  std::unordered_map<std::string, ir::Expr> loops_to_flat_var_map;
-  for (int idx = 0; idx < strides.size(); ++idx) {
-    if (strides[idx] == 1) {
-      // flat_i_to_loop_var.push_back(_iter);
-      loops_to_flat_var_map[loops[idx].As<ir::For>()->loop_var->name] = _iter;
-    } else {
-      // flat_i_to_loop_var.push_back(_iter / Expr(strides[idx]));
-      loops_to_flat_var_map[loops[idx].As<ir::For>()->loop_var->name] =
-          _iter / Expr(strides[idx]);
-      _iter = _iter % Expr(strides[idx]);
-    }
-  }
-
-  ir::FindBlocksVisitor visitor;
-  auto blocks = visitor(&last->body);
-  auto can_do_flat = [](const std::vector<Expr>& indexs,
-                        const std::vector<Var>& loop_vars) {
-    if (indexs.size() != loop_vars.size()) {
-      return false;
-    }
-
-    for (int idx = 0; idx < indexs.size(); ++idx) {
-      if (!indexs[idx].as_var()) {
-        return false;
-      } else {
-        auto var = indexs[idx].as_var_ref();
-        if (var->name != loop_vars[idx]->name) {
-          return false;
-        }
-      }
-    }
-    return true;
-  };
-
-  // change blocks iter value/iter var
-  for (auto& block : blocks) {
-    auto block_realize = block.As<ir::ScheduleBlockRealize>();
-    auto schedule_block = block_realize->schedule_block.As<ir::ScheduleBlock>();
-
-    // checkout loops in orders.
-    std::vector<std::string> var_names = {};
-    CHECK_GE(block_realize->iter_values.size(), loop_vars.size())
-        << "the number of iter bind values must be more than loop vars!";
-    for (int idx = 0; idx < block_realize->iter_values.size(); ++idx) {
-      auto& iter = block_realize->iter_values[idx];
-      if (iter.is_var()) {
-        CHECK_EQ(iter.as_var_ref()->name, loop_vars[idx]->name)
-            << "loops is not the same order with tensor!";
-      } else {
-        CHECK(iter.As<IntImm>()) << iter.node_type() << " is not IntImm";
-        CHECK_EQ(iter.as_int32(), 0);
-      }
-    }
-
-    auto exprs = ir::ir_utils::CollectIRNodesInOrder(
-        schedule_block->body,
-        [&](const Expr* x) { return x->As<ir::Store>() || x->As<ir::Load>(); });
-    // reverse exprs from last to first.
-    std::reverse(std::begin(exprs), std::end(exprs));
-
-    std::vector<ir::Var> var_to_replace;
-    std::vector<ir::Expr> flat_i_to_loop_var;
-    // if iter var is more than flat i to loop, there exist dim = 1.
-    for (int idx = 0; idx < block_realize->iter_values.size(); ++idx) {
-      if (block_realize->iter_values[idx].is_var()) {
-        var_to_replace.push_back(schedule_block->iter_vars[idx]);
-        auto var_name = block_realize->iter_values[idx].as_var_ref()->name;
-        CHECK(loops_to_flat_var_map.count(var_name))
-            << "Can't find var name : " << var_name;
-        flat_i_to_loop_var.push_back(loops_to_flat_var_map[var_name]);
-      } else {
-        CHECK_EQ(block_realize->iter_values[idx].as_int32(), 0);
-        // insert var -> 0, to replace var to 0.
-        var_to_replace.push_back(schedule_block->iter_vars[idx]);
-        flat_i_to_loop_var.push_back(Expr(0));
-      }
-    }
-    CHECK_EQ(var_to_replace.size(), flat_i_to_loop_var.size());
-
-    for (auto expr : exprs) {
-      if (expr.As<ir::Store>()) {
-        auto store = expr.As<ir::Store>();
-        if (store->is_addr_tensor()) {
-          auto t = store->tensor.as_tensor_ref();
-          CHECK(!t->reduce_axis.size());
-          auto tsize = std::accumulate(t->shape.begin(),
-                                       t->shape.end(),
-                                       1,
-                                       [](const int sum, const Expr& expr) {
-                                         return sum * expr.as_int32();
-                                       });
-          if ((!flat_tensor &&
-               !can_do_flat(store->indices, schedule_block->iter_vars)) ||
-              extent != tsize) {
-            // just replace indexs
-            for (auto& indice : store->indices) {
-              if (!indice.is_var()) {
-                continue;
-              }
-              ReplaceExpr(&indice, var_to_replace, flat_i_to_loop_var);
-            }
-            // compute index and flat tensor.
-            store->indices = {store->index()};
-            continue;
-          }
-          // update var and shape
-          store->indices = {Expr(_var)};
-        }
-      } else {
-        auto load = expr.As<ir::Load>();
-        if (load->is_addr_tensor()) {
-          auto t = load->tensor.as_tensor_ref();
-          CHECK(!t->reduce_axis.size());
-          auto tsize = std::accumulate(t->shape.begin(),
-                                       t->shape.end(),
-                                       1,
-                                       [](const int sum, const Expr& expr) {
-                                         return sum * expr.as_int32();
-                                       });
-          if ((!flat_tensor &&
-               !can_do_flat(load->indices, schedule_block->iter_vars)) ||
-              extent != tsize) {
-            // just replace indexs
-            for (auto& indice : load->indices) {
-              if (!indice.is_var()) {
-                continue;
-              }
-              ReplaceExpr(&indice, var_to_replace, flat_i_to_loop_var);
-            }
-            // compute index and flat tensor.
-            load->indices = {load->index()};
-            continue;
-          }
-          // update var and shape
-          load->indices = {Expr(_var)};
-        }
-      }
-    }
-    ReplaceExpr(&schedule_block->body, var_to_replace, flat_i_to_loop_var);
-
-    // update iter values
-    auto iter = ir::Expr(var);
-    block_realize->iter_values = {iter};
-
-    // update iter_vars
-    schedule_block->iter_vars = {_var};
-    CHECK_EQ(block_realize->iter_values.size(),
-             schedule_block->iter_vars.size());
-  }
-
-  this->Replace(loops[0], loop);
-  VLOG(4) << "After FlattenLoops, ir is:\n" << loop;
-}
-
-void StScheduleImpl::CopyTransformAndLoopInfo(
-    const std::string& block_name, const std::string& block_target_name) {
-  auto block = this->GetBlock(block_name);
-  auto block_target = this->GetBlock(block_target_name);
-  this->CopyTransformAndLoopInfo(block, block_target);
-}
-
-void StScheduleImpl::CopyTransformAndLoopInfo(const Expr& block,
-                                              const Expr& block_target) {
-  CHECK(block.As<ir::ScheduleBlockRealize>());
-  CHECK(block_target.As<ir::ScheduleBlockRealize>());
-  auto exprs = this->GetModule().GetExprs();
-  CHECK_EQ(exprs.size(), 1U);
-  auto expr = exprs[0];
-  auto vars = block.As<ir::ScheduleBlockRealize>()
-                  ->schedule_block.As<ir::ScheduleBlock>()
-                  ->iter_vars;
-  auto vars_target = block_target.As<ir::ScheduleBlockRealize>()
-                         ->schedule_block.As<ir::ScheduleBlock>()
-                         ->iter_vars;
-  auto old_iter_values = block.As<ir::ScheduleBlockRealize>()->iter_values;
-  auto iter_values_target =
-      block_target.As<ir::ScheduleBlockRealize>()->iter_values;
-  std::vector<Expr> new_iter_values;
-  for (int i = 0; i < vars.size() && i < vars_target.size(); ++i) {
-    CHECK(vars[i]->upper_bound.defined() &&
-          vars_target[i]->upper_bound.defined());
-    if (vars[i]->upper_bound.is_constant() &&
-        vars_target[i]->upper_bound.is_constant() &&
-        vars[i]->upper_bound.get_constant() ==
-            vars_target[i]->upper_bound.get_constant() &&
-        !vars[i]->is_reduce_axis && !vars_target[i]->is_reduce_axis) {
-      new_iter_values.push_back(iter_values_target[i]);
-      VLOG(3) << "new_iter_values.push_back " << iter_values_target[i];
-    } else {
-      break;
-    }
-  }
-
-  if (new_iter_values.empty())
-    LOG(FATAL) << "Cannot CopyTransformAndLoopInfo since shape[0] of source "
-                  "and target is not equal! "
-               << vars[0]->upper_bound << " v.s "
-               << vars_target[0]->upper_bound;
-
-  int changed_loop_num = new_iter_values.size();
-  std::set<std::string> used_target_loop_vars;
-  for (auto& iter_val : new_iter_values) {
-    auto find_partial_loop =
-        ir::ir_utils::CollectIRNodesWithoutTensor(iter_val, [&](const Expr* x) {
-          if (x->as_var()) used_target_loop_vars.insert(x->as_var_ref()->name);
-          return x->as_var();
-        });
-  }
-  CHECK(!used_target_loop_vars.empty());
-  std::vector<Expr> used_target_loops;
-  auto expr_copy = ir::ir_utils::IRCopy(expr);
-  for (auto& var : used_target_loop_vars) {
-    auto find_loop_var = ir::ir_utils::CollectIRNodesWithoutTensor(
-        expr_copy,
-        [&](const Expr* x) {
-          return x->As<ir::For>() && x->As<ir::For>()->loop_var->name == var &&
-                 Contains(*x, block_target);
-        },
-        true);
-    CHECK_EQ(find_loop_var.size(), 1U);
-    used_target_loops.push_back(*find_loop_var.begin());
-    VLOG(3) << "used_target_loops push_back " << used_target_loops.back();
-  }
-  std::sort(
-      used_target_loops.begin(), used_target_loops.end(), [&](Expr i, Expr j) {
-        return (utils::GetStreamCnt(i).size() > utils::GetStreamCnt(j).size());
-      });
-  for (int i = new_iter_values.size(); i < old_iter_values.size(); ++i) {
-    CHECK(old_iter_values[i].as_var());
-    new_iter_values.push_back(old_iter_values[i]);
-  }
-  Expr new_loop;
-  VLOG(3) << "changed_loop_num is : " << changed_loop_num;
-  VLOG(3) << "old_iter_values.size() is : " << old_iter_values.size();
-  if (changed_loop_num >= static_cast<int>(old_iter_values.size())) {
-    new_loop = ir::ir_utils::IRCopy(block);
-    new_loop.As<ir::ScheduleBlockRealize>()->iter_values = new_iter_values;
-  } else {
-    CHECK(old_iter_values[changed_loop_num].as_var());
-    auto old_var = old_iter_values[changed_loop_num].as_var_ref();
-    auto find_partial_loop = ir::ir_utils::CollectIRNodesWithoutTensor(
-        expr,
-        [&](const Expr* x) {
-          return x->As<ir::For>() &&
-                 x->As<ir::For>()->loop_var->name == old_var->name &&
-                 Contains(*x, block);
-        },
-        true);
-    CHECK_EQ(find_partial_loop.size(), 1U);
-    new_loop = ir::ir_utils::IRCopy(*find_partial_loop.begin());
-    auto find_schedule_block = ir::ir_utils::CollectIRNodesWithoutTensor(
-        new_loop,
-        [&](const Expr* x) { return x->As<ir::ScheduleBlockRealize>(); },
-        true);
-    CHECK_EQ(find_schedule_block.size(), 1U);
-    Expr sch_block = (*find_schedule_block.begin());
-    sch_block.As<ir::ScheduleBlockRealize>()->iter_values = new_iter_values;
-  }
-  VLOG(3) << "new_loop is : " << new_loop;
-  CHECK(!used_target_loops.empty());
-  Expr res;
-  if (used_target_loops.size() == 1) {
-    auto for_loop = used_target_loops[0].As<ir::For>();
-    res = For::Make(for_loop->loop_var,
-                    for_loop->min,
-                    for_loop->extent,
-                    for_loop->for_type(),
-                    for_loop->device_api,
-                    new_loop,
-                    for_loop->vectorize_info(),
-                    for_loop->bind_info());
-  } else {
-    Expr outer_loop = used_target_loops.front();
-    Expr inner_loop = used_target_loops.back();
-    inner_loop.As<ir::For>()->body = Block::Make({new_loop});
-    res = outer_loop;
-  }
-  VLOG(3) << "res is : " << res;
-  std::vector<Expr> all_loops = this->GetLoops(block);
-  CHECK(!all_loops.empty());
-  this->Replace(all_loops[0], res);
-}
-
-std::vector<Expr> StScheduleImpl::SamplePerfectTile(
-    utils::LinearRandomEngine::StateType* rand_seed,
-    const Expr& loop,
-    int n,
-    int max_innermost_factor) {
-  CHECK(loop.As<ir::For>())
-      << "Expr param of SamplePerfectTile should be a For loop";
-  CHECK_GE(n, 2) << "The number of tile factors should be at least 2";
-  CHECK_GE(max_innermost_factor, 1)
-      << "The max innermost factor should be at least 1";
-  CHECK(common::is_zero(loop.As<ir::For>()->min))
-      << "The For loop should start from 0";
-  int loop_extent = GetLoopExtent(loop);
-  std::vector<int> innermost_factors;
-  for (int i = max_innermost_factor; i >= 1; --i) {
-    if (loop_extent % i == 0) {
-      innermost_factors.push_back(i);
-    }
-  }
-  CHECK(!innermost_factors.empty()) << "No innermost factor found";
-  int innermost_factor = innermost_factors[utils::SampleUniformInt(
-      0, innermost_factors.size(), rand_seed)];
-  auto result = SampleTile(rand_seed, n - 1, loop_extent / innermost_factor);
-  std::vector<Expr> result_expr;
-  for (auto& factor : result) {
-    result_expr.push_back(Expr(factor));
-  }
-  result_expr.push_back(Expr(innermost_factor));
-  return result_expr;
-}
-
-Expr StScheduleImpl::SampleCategorical(
-    utils::LinearRandomEngine::StateType* rand_seed,
-    const std::vector<int>& candidates,
-    const std::vector<float>& probs) {
-  // check two sizes
-  CHECK_EQ(candidates.size(), probs.size())
-      << "candidates and probs must have same size.";
-  int seed_idx = utils::SampleDiscreteFromDistribution(probs, rand_seed);
-  auto result = candidates[seed_idx];
-  Expr result_expr(result);
-  return result_expr;
-}
 
 IRSchedule::IRSchedule() {}
